refactor(peagen): rename login_core to publickey_core
